Add environment specific configuration
Add .env files for environment specific configurations. The content of
these env files will later be added as environment variables of the
running system (e.g. Docker container).

For local development a program argument "dev" could be used. When this
argument exists, the .env.dev file is read at the server start and the
config is based directly on this file and not on environment variables.